Add error codes to unix scripts + better response processing

+++ scripts/unix/move.sh
@@ -4,4 +4,8 @@
 
 item="$1"
 destination="$2"
+if [[ -f "$destination" || -d "$destination" ]]
+then
+  exit 253
+fi
 mv "$item" "$destination"
